func-prog/recursion: correct solution for Super Digit

+++ functional-programming/recursion/super-digit.scala
@@ -1,15 +1,15 @@
 object Solution {
-    def digitSum(num: Int): Int = num.toString.toList.map(_.asDigit).sum
-    
-    def superDigit(num: Int): Int = num match {
+    def digitSum(num: BigInt): BigInt = num.toString.toList.map(_.asDigit).sum
+
+    def superDigit(num: BigInt): BigInt = num match {
         case n if n > 9 => superDigit(digitSum(n))
         case n => n
     }
 
     def main(args: Array[String]) {
         val sc = new java.util.Scanner (System.in)
-        val toRepeat = sc.nextInt()
-        val multip = sc.nextInt()
+        val toRepeat = BigInt(sc.nextBigInteger())
+        val multip = BigInt(sc.nextBigInteger())
         val in = superDigit(toRepeat) * multip
         println(superDigit(in))
     }
